Too bright glass ball.

+++ RRT/include/Camera.h
@@ -8,16 +8,16 @@ class Camera
 {
 public:
     Camera(const Vector3& center, const Vector3& gaze, const Vector3& vup,
-           float aperture, float distance, float u0, float v0, float u1, float v1);
+           double aperture, double distance, double u0, double v0, double u1, double v1);
     virtual ~Camera();
 
-    Ray getRay(float a, float b, float ra, float rb);
+    Ray getRay(double a, double b, double ra, double rb);
 
 private:
     Vector3 _center, _gaze;
     Vector3 _u, _v, _w;
     Vector3 _o, _a, _b;
-    float _radius;
+    double _radius;
 };
 
 #endif // CAMERA_H
