+++ RRT/include/TextureChess.h
@@ -6,14 +6,14 @@
 class TextureChess : public Texture
 {
 public:
-    TextureChess(const Color& color1, const Color& color2, float period);
+    TextureChess(const Color& color1, const Color& color2, double period);
     virtual ~TextureChess();
 
     virtual Color getColor(const Vector2& uv) const override final;
 
 private:
     Color _color1, _color2;
-    float _period;
+    double _period;
 };
 
 #endif // TEXTURECHESS_H
