+++ RRT/include/Vector2.h
@@ -7,62 +7,62 @@
 class Vector2
 {
 public:
-    Vector2(float x = 0.0f, float y = 0.0f);
+    Vector2(double x = 0.0, double y = 0.0);
     virtual ~Vector2();
 
-    float x() const;
-    float y() const;
-    float operator[](int index) const;
+    double x() const;
+    double y() const;
+    double operator[](int index) const;
 
-    void setX(float x);
-    void setY(float y);
-    void set(float x, float y);
+    void setX(double x);
+    void setY(double y);
+    void set(double x, double y);
 
     friend Vector2 operator+(const Vector2& a, const Vector2& b);
     friend Vector2 operator-(const Vector2& a, const Vector2& b);
     friend Vector2 operator*(const Vector2& a, const Vector2& b);
-    friend Vector2 operator*(const Vector2& a, float b);
-    friend Vector2 operator*(float a, const Vector2& b);
+    friend Vector2 operator*(const Vector2& a, double b);
+    friend Vector2 operator*(double a, const Vector2& b);
     friend Vector2 operator/(const Vector2& a, const Vector2& b);
-    friend Vector2 operator/(const Vector2& a, float b);
+    friend Vector2 operator/(const Vector2& a, double b);
 
-    friend float dot(const Vector2& a, const Vector2& b);
+    friend double dot(const Vector2& a, const Vector2& b);
 
-    float length() const;
+    double length() const;
     Vector2 norm() const;
 
     friend std::ostream& operator<<(std::ostream& out, const Vector2& vec);
 
 private:
-    float _d[2];
+    double _d[2];
 };
 
-inline float Vector2::x() const
+inline double Vector2::x() const
 {
     return _d[0];
 }
 
-inline float Vector2::y() const
+inline double Vector2::y() const
 {
     return _d[1];
 }
 
-inline float Vector2::operator[](int index) const
+inline double Vector2::operator[](int index) const
 {
     return _d[index];
 }
 
-inline void Vector2::setX(float x)
+inline void Vector2::setX(double x)
 {
     _d[0] = x;
 }
 
-inline void Vector2::setY(float y)
+inline void Vector2::setY(double y)
 {
     _d[1] = y;
 }
 
-inline void Vector2::set(float x, float y)
+inline void Vector2::set(double x, double y)
 {
     setX(x);
     setY(y);
@@ -83,12 +83,12 @@ inline Vector2 operator*(const Vector2& a, const Vector2& b)
     return Vector2(a.x() * b.x(), a.y() * b.y());
 }
 
-inline Vector2 operator*(const Vector2& a, float b)
+inline Vector2 operator*(const Vector2& a, double b)
 {
     return Vector2(a.x() * b, a.y() * b);
 }
 
-inline Vector2 operator*(float a, const Vector2& b)
+inline Vector2 operator*(double a, const Vector2& b)
 {
     return Vector2(a * b.x(), a * b.y());
 }
@@ -98,17 +98,17 @@ inline Vector2 operator/(const Vector2& a, const Vector2& b)
     return Vector2(a.x() / b.x(), a.y() / b.y());
 }
 
-inline Vector2 operator/(const Vector2& a, float b)
+inline Vector2 operator/(const Vector2& a, double b)
 {
     return Vector2(a.x() / b, a.y() / b);
 }
 
-inline float dot(const Vector2& a, const Vector2& b)
+inline double dot(const Vector2& a, const Vector2& b)
 {
     return a.x()*b.x() + a.y()*b.y();
 }
 
-inline float Vector2::length() const
+inline double Vector2::length() const
 {
     return sqrt(dot(*this, *this));
 }
