+++ RRT/include/Vector3.h
@@ -7,76 +7,76 @@
 class Vector3
 {
 public:
-    Vector3(float x = 0.0f, float y = 0.0f, float z = 0.0f);
+    Vector3(double x = 0.0, double y = 0.0, double z = 0.0);
     virtual ~Vector3();
 
-    float x() const;
-    float y() const;
-    float z() const;
-    float operator[](int index) const;
+    double x() const;
+    double y() const;
+    double z() const;
+    double operator[](int index) const;
 
-    void setX(float x);
-    void setY(float y);
-    void setZ(float z);
-    void set(float x, float y, float z);
+    void setX(double x);
+    void setY(double y);
+    void setZ(double z);
+    void set(double x, double y, double z);
 
     friend Vector3 operator+(const Vector3& a, const Vector3& b);
     friend Vector3 operator-(const Vector3& a, const Vector3& b);
     friend Vector3 operator-(const Vector3& a);
     friend Vector3 operator*(const Vector3& a, const Vector3& b);
-    friend Vector3 operator*(const Vector3& a, float b);
-    friend Vector3 operator*(float a, const Vector3& b);
+    friend Vector3 operator*(const Vector3& a, double b);
+    friend Vector3 operator*(double a, const Vector3& b);
     friend Vector3 operator/(const Vector3& a, const Vector3& b);
-    friend Vector3 operator/(const Vector3& a, float b);
+    friend Vector3 operator/(const Vector3& a, double b);
 
-    friend float dot(const Vector3& a, const Vector3& b);
+    friend double dot(const Vector3& a, const Vector3& b);
     friend Vector3 cross(const Vector3& a, const Vector3& b);
 
-    float length() const;
+    double length() const;
     Vector3 norm() const;
 
     friend std::ostream& operator<<(std::ostream& out, const Vector3& vec);
 
 private:
-    float _d[3];
+    double _d[3];
 };
 
-inline float Vector3::x() const
+inline double Vector3::x() const
 {
     return _d[0];
 }
 
-inline float Vector3::y() const
+inline double Vector3::y() const
 {
     return _d[1];
 }
 
-inline float Vector3::z() const
+inline double Vector3::z() const
 {
     return _d[2];
 }
 
-inline float Vector3::operator[](int index) const
+inline double Vector3::operator[](int index) const
 {
     return _d[index];
 }
 
-inline void Vector3::setX(float x)
+inline void Vector3::setX(double x)
 {
     _d[0] = x;
 }
 
-inline void Vector3::setY(float y)
+inline void Vector3::setY(double y)
 {
     _d[1] = y;
 }
 
-inline void Vector3::setZ(float z)
+inline void Vector3::setZ(double z)
 {
     _d[2] = z;
 }
 
-inline void Vector3::set(float x, float y, float z)
+inline void Vector3::set(double x, double y, double z)
 {
     setX(x);
     setY(y);
@@ -103,12 +103,12 @@ inline Vector3 operator*(const Vector3& a, const Vector3& b)
     return Vector3(a.x() * b.x(), a.y() * b.y(), a.z() * b.z());
 }
 
-inline Vector3 operator*(const Vector3& a, float b)
+inline Vector3 operator*(const Vector3& a, double b)
 {
     return Vector3(a.x() * b, a.y() * b, a.z() * b);
 }
 
-inline Vector3 operator*(float a, const Vector3& b)
+inline Vector3 operator*(double a, const Vector3& b)
 {
     return Vector3(a * b.x(), a * b.y(), a * b.z());
 }
@@ -118,12 +118,12 @@ inline Vector3 operator/(const Vector3& a, const Vector3& b)
     return Vector3(a.x() / b.x(), a.y() / b.y(), a.z() / b.z());
 }
 
-inline Vector3 operator/(const Vector3& a, float b)
+inline Vector3 operator/(const Vector3& a, double b)
 {
     return Vector3(a.x() / b, a.y() / b, a.z() / b);
 }
 
-inline float dot(const Vector3& a, const Vector3& b)
+inline double dot(const Vector3& a, const Vector3& b)
 {
     return a.x()*b.x() + a.y()*b.y() + a.z()*b.z();
 }
@@ -135,7 +135,7 @@ inline Vector3 cross(const Vector3& a, const Vector3& b)
                    a.x()*b.y() - a.y()*b.x());
 }
 
-inline float Vector3::length() const
+inline double Vector3::length() const
 {
     return sqrt(dot(*this, *this));
 }
