+++ RRT/src/TextureChess.cpp
@@ -1,6 +1,6 @@
 #include "TextureChess.h"
 
-TextureChess::TextureChess(const Color& color1, const Color& color2, float period) :
+TextureChess::TextureChess(const Color& color1, const Color& color2, double period) :
     _color1(color1), _color2(color2), _period(period)
 {
 }
